DixonGPTWebAPI/requirements.txt
llama-cpp-python
boto3
fastapi
uvicorn
# The `server` extra includes Flask and `pydantic`